refactor: remove vuu-utils2 package
