+++ vuu-ui/packages/vuu-data-remote/package.json
@@ -19,7 +19,6 @@
   },
   "dependencies": {
     "@vuu-ui/vuu-filter-parser": "2.3.0",
-    "@vuu-ui/vuu-utils": "2.3.0",
-    "@vuu-ui/vuu-utils2": "2.3.0"
+    "@vuu-ui/vuu-utils": "2.3.0"
   }
 }
