+++ vuu-ui/packages/vuu-data-test/package.json
@@ -11,8 +11,7 @@
   },
   "dependencies": {
     "@vuu-ui/vuu-data-local": "2.3.0",
-    "@vuu-ui/vuu-utils": "2.3.0",
-    "@vuu-ui/vuu-utils2": "2.3.0"
+    "@vuu-ui/vuu-utils": "2.3.0"
   },
   "devDependencies": {
     "@vuu-ui/vuu-protocol-types": "2.3.0",
