+++ vuu-ui/packages/vuu-utils/src/edit-utils.ts
@@ -1,0 +1,6 @@
+import type { EditSessionMode } from "@vuu-ui/vuu-data-types";
+
+export const isInlineEditingSession = (
+  editSessionMode: EditSessionMode,
+): editSessionMode is "inline-all-rows" =>
+  editSessionMode.startsWith("inline");
