--- vuu-ui/packages/vuu-utils2/package.json
@@ -1,22 +1,0 @@
-{
-  "name": "@vuu-ui/vuu-utils2",
-  "version": "2.3.0",
-  "author": "heswell",
-  "main": "src/index.ts",
-  "license": "Apache-2.0",
-  "types": "src/index.ts",
-  "type": "module",
-  "scripts": {
-    "build:dev": "node ../../scripts/run-build.mjs",
-    "build": "node ../../scripts/run-build-rollup.mjs",
-    "test": "vitest run",
-    "type-defs": "node ../../scripts/build-type-defs.mjs"
-  },
-  "devDependencies": {
-    "@vuu-ui/vuu-data-types": "2.3.0"
-  },
-  "peerDependencies": {
-    "react": "19.2.3"
-  },
-  "sideEffects": false
-}
