--- vuu-ui/packages/vuu-utils2/src/context-definitions/DataProvider.tsx
@@ -1,20 +1,0 @@
-import { type ReactNode, useContext } from "react";
-import { DataContext, type DataContextProps } from "./DataContext";
-
-export const DataProvider = ({
-  children,
-  getServerAPI,
-  isLocalData = true,
-  VuuDataSource,
-}: Omit<DataContextProps, "isLocalData"> & {
-  children: ReactNode;
-  isLocalData?: boolean;
-}) => {
-  return (
-    <DataContext.Provider value={{ isLocalData, VuuDataSource, getServerAPI }}>
-      {children}
-    </DataContext.Provider>
-  );
-};
-
-export const useData = () => useContext(DataContext);
