--- vuu-ui/packages/vuu-utils2/src/index.ts
@@ -1,5 +1,0 @@
-/** Context declarations hosted in utils to minimize intra package dependencies */
-export { isInlineEditingSession } from "./edit-utils";
-export { StaleUpdateError } from "./errors";
-export { DataContext } from "./context-definitions/DataContext";
-export { DataProvider, useData } from "./context-definitions/DataProvider";
