--- vuu-ui/packages/vuu-utils2/tsconfig.json
@@ -1,6 +1,0 @@
-{
-"extends": "../../tsconfig.json",
-"compilerOptions":{
-    "composite": true
-},
-}
